Split item sheet classes
Splits the TorchbearerItemSheet class into TorchbearerItemSheet and
TorchbearerSpellSheet.
