+++ module/item/spell-sheet.js
@@ -1,0 +1,11 @@
+import { TorchbearerBaseItemSheet } from "./base-item-sheet.js";
+
+export class TorchbearerSpellSheet extends TorchbearerBaseItemSheet {
+  /** @override */
+  static get defaultOptions() {
+    return mergeObject(super.defaultOptions, {
+      classes: ["torchbearer", "sheet", "spell"],
+      template: "systems/torchbearer/templates/item/spell-sheet.html.hbs",
+    });
+  }
+}
